doc: script to generate cfg var lists from C code
Added a perl script that generates the list of config variables
 defined in a C file.
The script uses gcc to generate a syntax tree structure for an
entire file and then parses and uses the tree to locate
the config vars. definition, initialization and default values.

E.g.: dump_cfg_defs.pl --file ../../../cfg_core.c --grp core
=> ...
31. core.udp_mtu
        fallback to a congestion controlled protocol if send size
        exceeds udp_mtu.
        Default: 0.
        Range: 0 - 65535.
        Type: integer.
...

The config group (--grp) needs to be specified only if it cannot
be auto-detected (under normal circumstances it is auto-detected
from the type name of the variable holding the default values,
e.g.  cfg_group_core).
If the default values are set in the same .c file, they will be
added to the list.

Note: the script requires a patched version of the
GCC:TranslationUnit perl package. To get the patch use:
dump_cfg_defs.pl --patch