doc: script to generate rpc lists from C code
Added a perl script that generates the list of RPCs defined in a C
file. The list contains also the RPC documentation.
The script works by looking for the first rpc_export_t array in
the file that has an initializer and using it to get all the rpc
definitions. Then it looks for all the corresponding *doc
variables (char* arrays holding the help message and the
signature) and extracts the doc string from their initializer.

E.g.:
dump_rpcs.pl --file ../../../core_cmd.c
...
37. dst_blacklist.view
        dst blacklist dump in human-readable format.
...