doc: script to generate select lists from C code
Added a perl script that tries to generate the list of selects
defined in a C file.
The script works by looking for the first select_row_t array with
an initializer in the .c file. It then tries to generate the list
of all possible selects (but it still has some bugs, especially on
 matching params to a select part).

Note: there is no documentation generated besides the list
(the select format is not auto-documented).