sr-oob.cfg: Automatic reload of global attributes.
This patch creates a new attribute called gattr_timestamp. The
attribute is stored in global_attrs table and contains the timestamp
of the last modification of the contents of the table. It can be
updated by applications like serweb whenever they update any of the
attributes in that table.

In the configuration file sr-oob.cfg we check the value stored in that
attribute and if its value indicates that it has been modified within
last two minutes, the script initiates a reload of the global_attr
memory cache.

The mechanism is the same as for the domain_attrs table reload that
is present in the same route section.

This patch also contains an unrelated small fix that correct the path
to the root directory in lib/srdb2/schema/Makefile. The root path
there was incorrect as result of the merge.