Improved multihomed performance
The mhomed implementation works by calling a socket()/connect()/getsockname()/close()
to find out the interface in a multihomed system.
Because mhomed works only for UDP sockets, a major performance improvement is shown
if we use the same socket for multiple connect()s, thus completly removing the socket() call
and the close() call.

The CPU load(user+system) shown in a mhomed environment on a stateless router, with a call rate of 6000 calls/s
is 46.1% load in the case of this patched version, versus 63.54% load in the case of the original
version.