tm: use ruri for forking only if marked as "new"
The r-uri is now used for forking only if marked as "new".
After using it, it will be automatically marked as "consumed", so
that it won't be used in the future (unless re-marked as new by
changing the r-uri, dst_uri or path).
This will simplify a lot serial forking via the script
(only seturi(...); t_relay() is needed now, no messing with
append_branch() is required anymore, unless parallel forking is desired)