tls: s/tls_cfg/tls_domains_cfg
Internal variable and type name change: replaces tls_cfg with
tls_domains_cfg. This frees up tls_cfg for use with the runtime
configuration framework (by convention a module runtime
configuration is named <module_name>_cfg and it would be very
confusing to have an exception for the tls module).