raw sockets: udp send will use now raw sockets if enabled
A raw socket, that will be used for sending, is now initialized on
startup. The operation can succeed only if sr is started as root
or with CAP_NET_RAW.
If the initialization fails and the raw sockets are forced on from
the config (udp4_raw = 1), ser won't start. If the raw socket are
in auto mode (udp4_raw = -1), sr will start, but with the raw
socket send part disabled (normal udp send will be used).

If the raw socket part is not disabled at startup, it can be
turned anytime on/off via the runtime cfg. framework.

Note: the raw socket send is supported only over ipv4 and for now
only on linux.