dns cache: dns_cache_rec_pref config var added
The config variable, dns_cache_rec_pref, can be used
to set the DNS cache preference as follows:

	0 - do not check duplicates (default)
	1 - prefer old records
	2 - prefer new records
	3 - prefer records with longer lifetime

This variable is checked when a duplicated record is
tried to be inserted into the cache. For instance the
SRV query answer contains also an A record which is
already in the cache.
If the config var is not 0, then permanent entries
are always preferred.

Note: works only with #define CACHE_RELEVANT_RECS_ONLY
at the moment.