modules_k/pv : Fixed a bug in pv transformation (tobody transformation), that caused a segmentation fault.
The bug was caused by a incorrect error handling of the parse_to() method (defined in parser/parse_to.c).
The function returns a char* (pointer to the last character parsed), in case of error it updates an error status.
On incorrect to params, a segmentation fault happened after some time.(Because of a double pkg_free).