modules_k/presence and modules_k/rls Fixed parse_to() usage
In case of error parse_to returns a pointer to the last parsed character. This is not a strict
error check, as several other modules provide a relaxed way of parsing To header (the uri is still
parseable even if the TO headers are not).
(cherry picked from commit 53c179bb3042bb767c43cd1a87a2cb8ab0ea92b7)