cfg framework: cfg_set_now() supports the group instances
cfg_set_now() can write the variables within both
the default and the additional group instances.
If the additional variable is not explicitely set, then
the default value overwrites it.

Note: the group instance must exist before cfg_set_now() is called,
it is not created dynamicaly.

The function does not work yet before the config is shmized.