tls.cfg: use local file paths
Use local paths, relative to the source/compile directory
(./modules/tls/*). This allows starting with -A LOCAL_TEST_RUN
with one of the sip-router*.cfg without having to edit any path in
tls.cfg.
On install-cfg, the local path is changed to an absolute path
to the installed certificates/keys/ca list (see 0efe88e).