cfg_rpc: Support for cfg group instances added.
The RPC interface of the configuration framework is updated
to the latest core changes, i.e. the support for additional group
instances is introduced.

The syntax is backward compatible:
 - To set the default value:
     cfg.set_now_int/string group var value
     cfg.set_delayed_int/string group var value
 - To set a value within a specific instance:
     cfg.set_now_int/string group[id] var value
     cfg_set_delayed_int/string group[id] var value

"id" is the integer identifier of the group instance which needs to
be created with:
  cfg.add_group_inst group[id]

The instance can also be deleted with:
  cfg.del_group_inst group[id]

Note that the variables of a group instance take the default values
from the active configucation after the instance is created. When a
default value changes within a group, the group instance values are
also updated with the change unless the variable has been explicitly
set in the group instance.