ctl: rpc->scan does not immediately send out errors
When the rpc scan function cannot read a parameter because of
a type mismatch, the function only prepares the error reply but
does not send it out immediately. The scan can be retried with
different parameter specification which resets the prepared
error message.

This allows variable types of parameters of the same rpc function,
for example:

if (rpc->scan(c, "d", &i) == 1)
	/* int parameter */
else if (rpc->scan(c, "s", &ch) == 1)
	/* char* parameter */
else
	return /* error */