modules_k/rr Do not falsely recover after strict route if RURI matches the server
Check if the RURI contains an lr param. If it doesn't, than the request
doesn't come after a strict route server, as Kamailio always adds a lr parameter(always loose routes)
This fixes problems with re-SUBSCRIBEs where RURI identifies the resourse and can be the same as the initial
subscribe.