cfg framework: set the local config in PROC_INIT
The local configuration of the main process is temporary
set during initialization in child_init with rank==PROC_INIT,
which makes the configuration group instances available for
the modules.
The local configuration is reset back to NULL afterwards to make
sure that each child process updates its own config the usual way.