cfg framework: set the handle before creating the cfg group
The handle must be set before the config group is created
because the related functions save the value for future use.

The tcp and sctp config handles are not initialized to the default
config struction. This cased a crash because orig_handle was set to
the initial value which is uninitialized.