modules_k/sqlops: Move s.sql transformation to sqlops as sql.val and add sql.val.int, sql.val.str
s.sql was not really a string transformation as it handled null and integers
too. Move it to sqlops as sql.val (it was not yet in a released version) and
add 2 new transformations.

Summarizing:

sql.val: returns a valid SQL value. $null as NULL, int as number, string as
         quoted and escaped string
sql.val.int: returns a non-null integer ($null becomes 0)
sql.val.str: returns a non-null quoted and escaped string. ($null becomes '')