modules_k/nathelper: fix add_contact_alias() for contacts without angle brackets <>
When adding the brackets and the alias, the function suffered from the
infamous problem that the same part of a message can only be changed once.
Inserting 3 lumps created an offset problem resulting in the alias
parameter being added outside the brackets. This patch converts it to
adding just 2 non-overlapping lumps by combining the closing > with the
alias parameter.